Add unified reminders/notifications & DB migration reset
- Introduced unified notification system: reminders and notifications are now merged in the UI, with read/unread tracking for both.
- Added IsRead property to ReminderLog and API endpoints to mark reminders as read.
- Changed ProgressScore (GsdeanEvaluation) and ScientificScore (VpgsEvaluation) to int for consistency.
- Added validation for past dates and improved input validation for scores.
- Major DB migration reset: removed old migrations, added new initial migration with updated schema and seed data.
- Updated DTOs, services, and Blazor components to support new features and types.

+++ Shared/Dtos/GSDean/CreateGsdeanEvaluationDto.cs
@@ -21,7 +21,7 @@ namespace Shared.Dtos.GSDean
         [Required]
         public int CompletedHours { get; set; }
         public decimal Gpa { get; set; }
-        public decimal? ProgressScore { get; set; }
+        public int? ProgressScore { get; set; }
         [Required]
         public DateOnly ExpectedCompletionDate {  get; set; }
         [StringLength(1000)]
