+++ Shared/Dtos/VPGSEvaluation/VpgsEvaluationResponseDto.cs
@@ -10,7 +10,7 @@ namespace Shared.Dtos.VPGSEvaluation
     {
         public int VpgsevalId { get; set; }
         public int EvaluationId { get; set; }
-        public decimal ScientificScore { get; set; }
+        public int ScientificScore { get; set; }
         public int TAEmployeeId { get; set; }
         public string PeriodName { get; set; }
         public string StatusName { get; set; }
